feat: Add option to generate random string in lowercase
Update the `generate_random_string` function in `misc.rs` to include a new parameter `to_lowercase` which allows generating random strings in lowercase if set to `true`. This provides more flexibility in generating random strings based on the requirements of the calling code.

Update the usage of `generate_random_string` in `clone.rs` to pass `true` as the value of the `to_lowercase` parameter. This ensures that the generated random string is in lowercase when creating the dump file.

chore: Update dependencies in Cargo.toml

Update the version of `clap` to `4.5.7` and `clap_complete` to `4.5.6` in the `dependencies` section of `Cargo.toml`. This ensures that the latest stable versions of these dependencies are used in the project.

chore: Update dependencies in Cargo.toml

Update the version of `csv` to `1.3.0` in the `dependencies` section of `Cargo.toml`. This ensures that the latest stable version of the `csv` dependency is used in the project.

chore: Update dependencies in Cargo.toml

Update the version of `serde` to `1.0` and `serde_json` to `1.0.118` in the `dependencies` section of `Cargo.toml`. This ensures that the latest stable versions of these dependencies are used in the project.

chore: Update dependencies in Cargo.toml

Update the version of `tar` to `0.4.41` in the `dependencies` section of `Cargo.toml`. This ensures that the latest stable version of the `tar` dependency is used in the project.

chore: Update dependencies in Cargo.toml

Update the version of `tokio` to `1.38.0` in the `dependencies` section of `Cargo.toml`. This ensures that the latest stable version of the `tokio` dependency is used in the project.