feat: Add support for BSQL format in database cloning
Update the `clone_db` function in `clone.rs` to include the `SqlFileFormat` enum as a parameter. This allows specifying the format of the dump file when cloning the database. In this commit, the `SqlFileFormat::Bsql` format is used.

Update the `db_dump` and `db_restore` functions in `general.rs` to include the `dump_format` and `restore_format` parameters respectively. These parameters are used to specify the format of the dump file when performing database dump and restore operations. In this commit, the `SqlFileFormat::Bsql` format is used.

This change provides more flexibility in choosing the format of the dump file when cloning and restoring databases.